less warnings1

+++ cg_new/cg/matrix4f.h
@@ -10,35 +10,35 @@ class Matrix4f
 public:
     Matrix4f();
     Matrix4f InitIdentity();
-    Matrix4f InitScreenSpaceTransform(double halfWidth, double halfHeight);
-    Matrix4f InitTranslation(double x, double y, double z);
-    Matrix4f InitRotation(double x, double y, double z, double angle);
-    Matrix4f InitRotation(double x, double y, double z);
+    Matrix4f InitScreenSpaceTransform(float halfWidth, float halfHeight);
+    Matrix4f InitTranslation(float x, float y, float z);
+    Matrix4f InitRotation(float x, float y, float z, float angle);
+    Matrix4f InitRotation(float x, float y, float z);
 
     Matrix4f InitRotation(Vector4f forward, Vector4f up);
     Matrix4f InitRotation(Vector4f forward, Vector4f up, Vector4f right);
 
 
-    Matrix4f InitScale(double x, double y, double z);
-    Matrix4f InitPerspective(double fov, double aspectRatio,
-                             double zNear, double zFar);
-    Matrix4f InitOrthographic(double left, double right,
-                              double bottom, double top,
-                              double near, double far);
+    Matrix4f InitScale(float x, float y, float z);
+    Matrix4f InitPerspective(float fov, float aspectRatio,
+                             float zNear, float zFar);
+    Matrix4f InitOrthographic(float left, float right,
+                              float bottom, float top,
+                              float near, float far);
 
     Vector4f Transform(Vector4f r);
     Matrix4f Mul(Matrix4f r);
 
-    std::vector <std::vector <double>> GetMtx();
-    void SetMtx(std::vector <std::vector <double>> mtx);
+    std::vector <std::vector <float>> GetMtx();
+    void SetMtx(std::vector <std::vector <float>> mtx);
 
 
-    double Get(int x, int y);
-    void Set(int x, int y, double value);
+    float Get(int x, int y);
+    void Set(int x, int y, float value);
 
 
 private:
-    std::vector <std::vector <double>> _mtx;
+    std::vector <std::vector <float>> _mtx;
 
 };
